+++ deployment/environment/devenv/chain.go
@@ -2,6 +2,7 @@ package devenv
 
 import (
 	"context"
+	"encoding/json"
 	"errors"
 	"fmt"
 	"math/big"
@@ -12,7 +13,9 @@ import (
 	"time"
 
 	"github.com/aptos-labs/aptos-go-sdk"
+	"github.com/ethereum/go-ethereum"
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
+	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/gagliardetto/solana-go"
@@ -201,7 +204,7 @@ func NewChains(ctx context.Context, logger logger.Logger, configs []ChainConfig)
 					}
 					blockNumber = receipt.BlockNumber.Uint64()
 					if receipt.Status == 0 {
-						errReason, err := deployment.GetErrorReasonFromTx(ec, chainCfg.DeployerKey.From, tx, receipt)
+						errReason, err := getErrorReasonFromTx(ec, chainCfg.DeployerKey.From, tx, receipt)
 						if err == nil && errReason != "" {
 							return blockNumber, fmt.Errorf("tx %s reverted,error reason: %s chain %s", tx.Hash().Hex(), errReason, chainInfo.ChainName)
 						}
@@ -387,4 +390,49 @@ func (c *ChainConfig) SetAptosDeployerKey(keyString *string) error {
 
 	c.AptosDeployerKey = *aptosAccount
 	return nil
+}
+
+func getErrorReasonFromTx(client bind.ContractBackend, from common.Address, tx *types.Transaction, receipt *types.Receipt) (string, error) {
+	call := ethereum.CallMsg{
+		From:     from,
+		To:       tx.To(),
+		Data:     tx.Data(),
+		Value:    tx.Value(),
+		Gas:      tx.Gas(),
+		GasPrice: tx.GasPrice(),
+	}
+	_, callContractErr := client.CallContract(context.Background(), call, receipt.BlockNumber)
+	if callContractErr != nil {
+		errorReason, parsingErr := parseError(callContractErr)
+		// If we get no information from parsing the error, we return the original error from CallContract
+		if errorReason == "" {
+			return callContractErr.Error(), nil
+		}
+		// If the errorReason exists and we had no issues parsing it, we return it
+		if parsingErr == nil {
+			return errorReason, nil
+		}
+	}
+	return "", fmt.Errorf("tx %s reverted with no reason", tx.Hash().Hex())
+}
+
+func parseError(txError error) (string, error) {
+	b, err := json.Marshal(txError)
+	if err != nil {
+		return "", err
+	}
+	var callErr struct {
+		Code    int
+		Data    string `json:"data"`
+		Message string `json:"message"`
+	}
+	if json.Unmarshal(b, &callErr) != nil {
+		return "", err
+	}
+
+	if callErr.Data == "" && strings.Contains(callErr.Message, "missing trie node") {
+		return "", errors.New("please use an archive node")
+	}
+
+	return callErr.Data, nil
 }
